move cargo related files in models
